requirements.txt
exchangelib
tzlocal
python-dateutil
git+https://github.com/andylytical/pyexch
#git+https://github.com/andylytical/pyexch@write_events
